start working on per-material textures

+++ src/plover_int.h
@@ -3,4 +3,4 @@
 #include "plover/plover.h"
 
 // Read entire contents of file into buffer and return buffer
-void readFile(const char *path, u8 **buffer, u64 *bufferSize);
+void readFile(const char *path, u8 **buffer, u32 *bufferSize);
